Solución ejercicios semillero semana 1

+++ In-Silico/Semana 1/B.cpp
@@ -1,11 +1,16 @@
+/*
+ *  HackerRank - Warmup Challenges
+ *  Problem: Simple Array Sum
+ *  Level: Easy
+ */
 #include <bits/stdc++.h>
 using namespace std;
 
 int main() {
   int n; cin >> n;
-  long long ans = 0;
+  int ans = 0;
   for (int i = 0; i < n; i++) {
-    long long x; cin >> x;
+    int x; cin >> x;
     ans += x;
   }
   cout << ans << endl;
